Introduce executable documetation to test the deployment docs.
Rationale:
We would like to test the deployment of the microservices-demo on top of
multiple orchectrators. This requires some steps to actually set up the
infrastructure for an orchestrator, which is also described in the
documentation.

With this comment, we actually run the steps in the documentation to set
up and tear down the orchestrator.