Various improvements to the load_tests (#531)
* created a separate section for the load tests, to be consistent with the other instructions

* Locust has a slightly weird behaviour when using --num-request flag.
That flag represents the number of requests after which locust will stop
running tests. That means that if locust will run with 3 clients, each
client will get to go through (num-requests/3) of requests before
terminating. And since our tests require a a larger number of requests
to complete, we need to increase the overal number of reuqests to make
sure that each client has enough to go trough all the tests.

* Locust has a slightly weird behaviour when using --num-request flag.
That flag represents the number of requests after which locust will stop
running tests. That means that if locust will run with 3 clients, each
client will get to go through (nu
m-requests/3) of requests before terminating. And since our tests
require a a larger number of requests to complete, we need to increase
the overal number of reuqests to make sure that each client has enough
to go trough all the tests.