Convert FrontEnd service to LoadBalancer
The use of NodePort works great for testing in local but for deploying
to cloud providers like AWS or GKE, when deploying the front-end service
as a NodePort, the port does indeed get allocated but no public IP gets
allocated, making the user to manually create a firewall rule to allow
incoming traffic to that very specific port. This is quite cumbersome.
By converting the service to a LoadBalancer, providers like GKE will
pick the service up, allocate a public IP to it and route all the
traffice to the corresponding endpoint transparently without user
intervention.